fix: seat 도메인이 api 명세대로 동작하도록 로직 수정

+++ src/main/java/com/hongik/fe_sit_ival_server/seat/dto/Coordinate.java
@@ -1,3 +1,3 @@
 package com.hongik.fe_sit_ival_server.seat.dto;
 
-public record Coordinate(int horizon, int vertical) {}
+public record Coordinate(Long horizon, Long vertical) {}
